@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net10.0</TargetFramework>
     <OutputType>Exe</OutputType>
     <StartupObject>WingetRepoBrowserCoreTests.Program</StartupObject>
   </PropertyGroup>
@@ -8,6 +8,7 @@
     <ProjectReference Include="..\WingetRepoBrowserCore\WingetRepoBrowserCore.csproj" />
   </ItemGroup>
   <ItemGroup>
+    <PackageReference Include="LibGit2Sharp" Version="0.29.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk">
       <Version>17.13.0</Version>
     </PackageReference>
